+++ src/jsonrepair.test.ts
@@ -1,8 +1,8 @@
 import { strictEqual, deepStrictEqual, throws } from 'assert'
-import { jsonrepair } from './jsonrepair.js'
+import { jsonrepair } from './index.js'
 import { JSONRepairError } from './JSONRepairError.js'
 
-describe('jsonRepair', () => {
+describe('jsonrepair', () => {
   describe('parse valid JSON', () => {
     it('parse full JSON object', function () {
       const text = '{"a":2.3e100,"b":"str","c":null,"d":false,"e":[1,2,3]}'
